+++ nuxeo-drive-client/tests/common.py
@@ -10,7 +10,7 @@ from unittest import TestCase
 
 import nxdrive
 from nxdrive.client import LocalClient, RemoteDocumentClient, \
-    RemoteFileSystemClient, RestAPIClient
+    RemoteFileSystemClient
 from nxdrive.client.common import BaseClient
 from nxdrive.logging_config import configure, get_logger
 from nxdrive.osi import AbstractOSIntegration
@@ -166,17 +166,6 @@ class IntegrationTestCase(TestCase):
             self.version,
             password=self.password_2, base_folder=self.workspace,
             upload_tmp_dir=self.upload_tmp_dir)
-
-        self.remote_restapi_client_1 = RestAPIClient(
-            self.nuxeo_url, self.user_1, u'nxdrive-test-device-1',
-            self.version,
-            password=self.password_1
-        )
-        self.remote_restapi_client_2 = RestAPIClient(
-            self.nuxeo_url, self.user_2, u'nxdrive-test-device-2',
-            self.version,
-            password=self.password_2
-        )
 
         # File system client to be used to create remote test documents
         # and folders